Refactored folder structure for base layout files
